backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
python-dotenv==1.0.1
bcrypt==4.2.0
python-jose==3.3.0
pydantic==2.9.2
email-validator==2.2.0